+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update Cluster API and Kubeadm imports to v1beta2 and refactor code accordingly.
+
+### Added
+
+- Add new functions `GetAWSManagedControlPlane` and `IsAWSManagedControlPlaneConditionSet` to check conditions for `AWSManagedControlPlane` resources.
+
 ## [3.1.3] - 2026-02-17
 
 ### Changed